+++ src/main/kotlin/racing/infrastructure/ResultView.kt
@@ -1,4 +1,6 @@
-package racing.domain
+package racing.infrastructure
+
+import racing.domain.RaceResult
 
 fun interface ResultView {
     fun showResult(result: RaceResult)
